WV-58-04: delete previous version of the screen
